feat(03-final): use Eleventy Universal Filter instead
This was suggested in a comment on my blog post!

Using a data file to make `@11ty/eleventy-img` seemed hacky, so this is much
more pleasing.

I'm still hopeful that it would be possible to require `@11ty/eleventy-img`
directly in the `_includes/webc/img.webc` file instead, to keep all image
related concerns there.